Initialize scheme.uuid to a default generated UUID.
This was causing crashes when creating new color schemes from scratch,
since schemes that weren't initialized from a property list had a nil
uuid. In turn, when converting the scheme to a property list, it tried
to put the nil uuid in the property list, which of course is invalid.