Remove name field from QScheme.
Decided it was confusing having the scheme name in the options, so the
name is now derived from what you save the tmTheme/stTheme as.

In addition, uuid is no longer observed since it's just there for
convention's sake, since I haven't looked into when TM/TM2 want it
updated. Presumably any newly-saved version should have a new UUID,
which is easy enough (generate the UUID every time toPropertyList
is called), but tedious. Might remove it from the scheme as well.