Add an ifDefined method. Basically just map for NSObject.
Doesn't accept any arguments because I figure you'll have the receiver
in scope somehow. Might change it later. I'd consider this particular
interface unstable.

Since I plan on pulling a lot of code out of this project for later
stuff, it's a bit of a testing ground. At any rate, I won't commit the
test-uses of this to the project to avoid too much unnecessary crap.